+++ bugsnag-android-core/src/main/java/com/bugsnag/android/SessionStore.java
@@ -2,6 +2,7 @@ package com.bugsnag.android;
 
 import android.content.Context;
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import java.io.File;
 import java.util.Comparator;
@@ -32,9 +33,10 @@ class SessionStore extends FileStore<Session> {
         }
     };
 
-    SessionStore(@NonNull Configuration config, @NonNull Context appContext) {
+    SessionStore(@NonNull Configuration config, @NonNull Context appContext,
+                 @Nullable Delegate delegate) {
         super(config, appContext, "/bugsnag-sessions/",
-            128, SESSION_COMPARATOR);
+            128, SESSION_COMPARATOR, delegate);
     }
 
     @NonNull