+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## TBD
+
+* Report internal errors when serialization fails
+  [#581](https://github.com/bugsnag/bugsnag-android/pull/581)
+
+* Buffer IO when reading from cached error files, improving SDK performance
+  [#573](https://github.com/bugsnag/bugsnag-android/pull/573)
+
 ## 4.19.1 (2019-09-03)
 
 ### Bug fixes